+++ src/cart/index.ts
@@ -1,2 +1,3 @@
-export * from './models';
+export * from './models/cart.entity';
+export * from './models/cart-item.entity';
 export * from './services';
