+++ src/cart/models/cart-item.entity.ts
@@ -1,0 +1,18 @@
+import { Entity, Column, ManyToOne, PrimaryColumn, JoinColumn } from 'typeorm';
+import { CartEntity } from './cart.entity';
+
+@Entity('cart_items')
+export class CartItemEntity {
+  @PrimaryColumn({ type: 'uuid' })
+  cart_id: string;
+
+  @PrimaryColumn({ type: 'uuid' })
+  product_id: string;
+
+  @Column({ type: 'int', default: 1 })
+  count: number;
+
+  @ManyToOne(() => CartEntity, (cart) => cart.items, { onDelete: 'CASCADE' })
+  @JoinColumn({ name: 'cart_id' })
+  cart: CartEntity;
+}
