+++ src/cart/models/cart.entity.ts
@@ -1,0 +1,35 @@
+import {
+  Entity,
+  PrimaryGeneratedColumn,
+  Column,
+  CreateDateColumn,
+  UpdateDateColumn,
+  OneToMany,
+} from 'typeorm';
+import { CartItemEntity } from './cart-item.entity';
+
+export enum CartStatus {
+  OPEN = 'OPEN',
+  ORDERED = 'ORDERED',
+}
+
+@Entity('carts')
+export class CartEntity {
+  @PrimaryGeneratedColumn('uuid')
+  id: string;
+
+  @Column({ type: 'uuid', nullable: false })
+  user_id: string;
+
+  @CreateDateColumn({ type: 'timestamp' })
+  created_at: Date;
+
+  @UpdateDateColumn({ type: 'timestamp' })
+  updated_at: Date;
+
+  @Column({ type: 'enum', enum: CartStatus, default: CartStatus.OPEN })
+  status: CartStatus;
+
+  @OneToMany(() => CartItemEntity, (cartItem) => cartItem.cart)
+  items: CartItemEntity[];
+}
