--- src/database/entities/cart-item.entity.ts
@@ -1,20 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column, ManyToOne } from 'typeorm';
-import { Cart } from './cart.entity';
-
-@Entity('cart_items')
-export class CartItem {
-  @PrimaryGeneratedColumn('uuid')
-  id?: string;
-
-  @Column({ type: 'uuid' })
-  cart_id: string;
-
-  @Column({ type: 'uuid' })
-  product_id: string;
-
-  @Column({ type: 'int' })
-  count: number;
-
-  @ManyToOne(() => Cart, (cart) => cart.items)
-  cart?: Cart;
-}
