--- src/database/entities/cart.entity.ts
@@ -1,39 +1,0 @@
-import {
-  Entity,
-  PrimaryGeneratedColumn,
-  Column,
-  CreateDateColumn,
-  UpdateDateColumn,
-  OneToMany,
-} from 'typeorm';
-import { CartItem } from './cart-item.entity';
-
-export enum CartStatus {
-  OPEN = 'OPEN',
-  ORDERED = 'ORDERED',
-}
-
-@Entity('carts')
-export class Cart {
-  @PrimaryGeneratedColumn('uuid')
-  id: string;
-
-  @Column({ type: 'uuid' })
-  user_id: string;
-
-  @CreateDateColumn()
-  created_at: Date;
-
-  @UpdateDateColumn()
-  updated_at: Date;
-
-  @Column({
-    type: 'enum',
-    enum: CartStatus,
-    default: CartStatus.OPEN,
-  })
-  status: CartStatus;
-
-  @OneToMany(() => CartItem, (cartItem) => cartItem.cart)
-  items: CartItem[];
-}
